[2.0] Fix up G33, LPC1768 + SDCARD_SORT_ALPHA (#8250)
* Update Conditionals_post.h

* Add a cast to round() to convert to a unsigned int

Add's a cast to round() so that it will compile properly. round() returns a float which must be cast to a integer for the following % operation. Use a unsigned int as a negative index to an array is wrong. Should never be more than 255 points allowing us to use a 8 bit cast.

* Update G33.cpp