[2.0.x][HD44780] Remove unused include
The <binary.h> header is not available in STM32 toolchain
and is not used anywhere in ultralcd_common_HD44780.h.
If it is used anywhere in HD44780 support for other platforms,
it must be included in the corresponding .cpp file directly
and put under appropriate conditional compilation directives
for the platform requiring it.

As I was unable to find such code, I consider the file unused
and hence remove the inclusion.