Fix Ender-2 LCD contrast
Changed the default to overcome the changes in:
8254fd9ddcaae4280fbaa607717d4a9ea70b2df5
This commit exposed a bug because the Ender-2 was using a MINIPANEL lcd
and getting default values that were too low, so the text was too weak.
However, the panel should be MKS_MINI_12864, since it
explicitly sets the correct contrast without relying in the defaults.